rag-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-dotenv
langchain
langchain-groq
langchain-text-splitters
langchain-community
faiss-cpu
groq
openai
langchain-openai